Move spec folder to gem_rspec to avoid test complications
