+++ .rspec
@@ -1,1 +1,2 @@
 --color
+--default-path gem_rspec
