+++ gem_rspec/fake_template.md.erb
