+++ gem_rspec/install_generator_spec.rb
@@ -2,9 +2,11 @@ require 'spec_helper'
 require 'generator_spec'
 require 'generators/smashing_documentation/install_generator'
 RSpec.describe SmashingDocumentation::Generators::InstallGenerator, type: :generator do
+  let(:rails_helper) { "spec/rails_helper.rb" }
+  let(:spec_helper) { "spec/spec_helper.rb" }
+  let(:docs_template) { "smashing_docs/template.md" }
   context "when rspec is installed" do
     describe "#update_spec_helper" do
-      let(:rails_helper) { "spec/rails_helper.rb" }
       it "appends the SmashingDocs config to rails_helper.rb" do
         run_generator
         expect(File).to exist(rails_helper)
@@ -13,17 +15,15 @@ RSpec.describe SmashingDocumentation::Generators::InstallGenerator, type: :gener
     end
 
     describe "#update_spec_helper" do
-      let(:spec_helper) { "spec/spec_helper.rb" }
       it "appends test suite hooks inside RSpec.configure block in spec_helper.rb" do
         run_generator
         expect(File).to exist(spec_helper)
-        expect(File.read(spec_helper)).to include("config.after(:each)")
+        expect(File.read(spec_helper)).to include("config.after(:each,")
         expect(File.read(spec_helper)).to include("config.after(:suite)")
       end
     end
 
     describe "#generate_docs_template" do
-      let(:docs_template) { "smashing_docs/template.md" }
       it "generates the default docs template" do
         run_generator
         expect(File).to exist(docs_template)
@@ -32,12 +32,14 @@ RSpec.describe SmashingDocumentation::Generators::InstallGenerator, type: :gener
       end
     end
   end
-  context "when RSpec is not installed" do
-    it "gives the user an error" do
-      File.rename 'spec', 's'
+  context "when rspec is not installed" do
+    it "does not install smashing_docs" do
+      File.rename('spec', 's') if Dir.exists?('spec')
       run_generator
-      expect(STDOUT).to receive(:puts).with("Please set up RSpec")
-      File.rename 's', 'spec'
+      expect(File).to_not exist(rails_helper)
+      File.rename('s', 'spec') if Dir.exists?('s')
+
+      # STDOUT is not tested here because it is suppressed in generator tests
     end
   end
 end
