+++ gem_rspec/rails_helper.rb
@@ -1,0 +1,5 @@
+SmashingDocs.config do |c|
+  c.template_file = 'smashing_docs/template.md'
+  c.output_file   = 'smashing_docs/api_docs.md'
+  c.run_all       = true
+end
